requirements.txt
aiofiles==0.3.2
backpack==0.1
blinker==1.4
cleo==0.6.1
click==6.7
Faker==0.7.18
gunicorn==19.7.1
httptools==0.0.10
inflection==0.3.1
lazy-object-proxy==1.3.1
mysqlclient==1.3.12
orator==0.9.7
pastel==0.1.0
pendulum==1.3.2
psycopg2==2.7.3.2
pyaml==16.12.2
Pygments==2.2.0
pylev==1.3.0
python-dateutil==2.6.1
python-dotenv==0.7.1
pytz==2017.3
pytzdata==2017.3.1
PyYAML==3.12
sanic==0.7.0
sanic-gunicorn==0.1.2
simplejson==3.13.2
six==1.11.0
tzlocal==1.5.1
ujson==1.35
uvloop==0.9.1
websockets==4.0.1
wrapt==1.10.11
